Enables BigUint/BigInt support in Wolfram Language
This commit refactors the Wolfram Language interface to use
strings for representing BigUint and BigInt values. This allows
arbitrary-precision integers to be passed to and from Wolfram
Language functions.

The `max_steps` parameter is also changed to `u64` to allow for a
larger number of steps in the Turing machine simulation.

The wolfram-library-link dependency is updated.